+++ SOSS/soss_tfit/core/base.py
@@ -14,9 +14,9 @@ Created on 2022-04-11
 # Define variables
 # =============================================================================
 __NAME__: str = 'base.py'
-__version__: str = '0.1'
-__date__: str = '2022-04-11'
-__authors__: str = 'Neil Cook, Jason Rowe'
+__version__: str = '0.4'
+__date__: str = '2022-04-13'
+__authors__: str = 'Neil Cook, Jason Rowe, David Lafreniere, Loic Albert'
 __package__: str = 'soss_tfit'