+++ SOSS/soss_tfit/science/general.py
@@ -15,7 +15,6 @@ from typing import Any, Dict, List, Optional, Type, Union
 
 from soss_tfit.core import base
 from soss_tfit.core import base_classes
-from soss_tfit.core import io
 
 # =============================================================================
 # Define variables
@@ -33,7 +32,7 @@ QUANTITIES = ['WAVELENGTH', 'FLUX', 'FLUX_ERROR']
 TQUANTITIES = ['WAVELENGTH', 'FLUX', 'FLUX_ERROR', 'TIME']
 
 # =============================================================================
-# Define functions
+# Define classes
 # =============================================================================
 class InputData:
     """
@@ -73,8 +72,7 @@ class InputData:
         # set orders
         self.orders = list(params['ORDERS'])
         # convert data into a stack
-        self.spec = io.stack_multi_spec(data, self.orders,
-                                        quantities=QUANTITIES)
+        self.spec = stack_multi_spec(data, self.orders, quantities=QUANTITIES)
         # storage for number of pixels in spectral dimensions
         self.n_wav = dict()
         # set number of pixels in each order
@@ -334,6 +332,56 @@ class InputData:
         # return the parameter dictionary
         return params
 
+
+# =============================================================================
+# Define functions
+# =============================================================================
+def stack_multi_spec(multi_spec, use_orders: List[int],
+                     quantities: List[str]) -> Dict[int, Dict[str, np.ndarray]]:
+    """
+    Convert a jwst multi_spec into a dictionary
+
+    all_spec[order][quantity]
+
+    where order is [1, 2, 3] etc
+    quantity is ['WAVELENGTH', 'FLUX', 'FLUX_ERROR', 'TIME', 'BIN_LIMITS']
+
+    :param multi_spec:
+    :param use_orders:
+    :param quantities:
+    :return:
+    """
+    # define a container for the stack
+    all_spec = dict()
+    # -------------------------------------------------------------------------
+    # loop around orders
+    for sp_ord in use_orders:
+        # set up dictionary per order for each quantity
+        all_spec[sp_ord] = dict()
+        # loop around quantities
+        for quantity in quantities:
+            all_spec[sp_ord][quantity] = []
+    # -------------------------------------------------------------------------
+    # loop around orders and get data
+    for spec in multi_spec.spec:
+        # get the spectral order number
+        sp_ord = spec.spectral_order
+        # skip orders not wanted
+        if sp_ord not in use_orders:
+            continue
+        # load values for each quantity
+        for quantity in quantities:
+            # only add quantities
+            if quantity in spec.spec_table.columns.names:
+                all_spec[sp_ord][quantity].append(spec.spec_table[quantity])
+    # -------------------------------------------------------------------------
+    # convert list to array
+    for sp_ord in all_spec:
+        for key in all_spec[sp_ord]:
+            all_spec[sp_ord][key] = np.array(all_spec[sp_ord][key])
+    # -------------------------------------------------------------------------
+    # return the all spectrum dictionary
+    return all_spec
 
 # =============================================================================
 # Start of code